Add files via upload
CredX is a leading credit card provider that gets thousands of credit card applicants every year. But in the past few years, it has experienced an increase in credit loss. The CEO believes that the best strategy to mitigate credit risk is to ‘acquire the right customers’.

We have helped CredX to identify the right set of customers by building an application score card using GLM,Random Forest,SMOTE+Random Forest,Logistic Regression.

We finally came to the conclusion that the model will automatically approve 62% of the applications and reject the remaining 38% of the customers, there by significantly improving the speed and accuracy of the transactions.